+++ app/Http/Requests/Statistics/SpaceSizeVariantRequest.php
@@ -9,14 +9,14 @@ use App\Contracts\Models\AbstractAlbum;
 use App\Http\Requests\BaseApiRequest;
 use App\Http\Requests\Traits\HasAbstractAlbumTrait;
 use App\Http\Requests\Traits\HasOwnerIdTrait;
-use App\Models\Configs;
+use App\Models\User;
 use App\Policies\AlbumPolicy;
-use App\Policies\SettingsPolicy;
+use App\Policies\UserPolicy;
 use App\Rules\RandomIDRule;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
 
-class SpacePerAlbumRequest extends BaseApiRequest implements HasAbstractAlbum, HasOwnerId
+class SpaceSizeVariantRequest extends BaseApiRequest implements HasAbstractAlbum, HasOwnerId
 {
 	use HasAbstractAlbumTrait;
 	use HasOwnerIdTrait;
@@ -27,7 +27,7 @@ class SpacePerAlbumRequest extends BaseApiRequest implements HasAbstractAlbum, H
 	public function authorize(): bool
 	{
 		if ($this->album === null) {
-			return Gate::check(SettingsPolicy::CAN_SEE_STATISTICS, [Configs::class]);
+			return Gate::check(UserPolicy::CAN_EDIT, [User::class]);
 		}
 
 		return Gate::check(AlbumPolicy::CAN_ACCESS, [AbstractAlbum::class, $this->album]);
