+++ resources/js/components/drawers/AlbumStatistics.vue
@@ -15,6 +15,8 @@
 				</template>
 			</Card>
 
+			<SizeVariantMeter v-if="props.config.is_model_album" :album-id="props.album.id" />
+
 			<DataTable
 				v-if="photosData.lens.length > 1"
 				:value="photosData.lens"
@@ -70,7 +72,6 @@
 import StatisticsService from "@/services/statistics-service";
 import { useLycheeStateStore } from "@/stores/LycheeState";
 import { sizeToUnit } from "@/utils/StatsSizeVariantToColours";
-import { TotalAlbum } from "@/views/Statistics.vue";
 import { storeToRefs } from "pinia";
 import Card from "primevue/card";
 import Column from "primevue/column";
@@ -78,25 +79,14 @@ import DataTable from "primevue/datatable";
 import Panel from "primevue/panel";
 import { computed, Ref, ref } from "vue";
 import { Collapse } from "vue-collapsed";
+import { TotalAlbum } from "../statistics/TotalCard.vue";
+import { useAlbumsStatistics } from "@/composables/album/albumStatistics";
+import SizeVariantMeter from "../statistics/SizeVariantMeter.vue";
 
 const lycheeStore = useLycheeStateStore();
 const { are_nsfw_visible } = storeToRefs(lycheeStore);
 
 const areStatisticsOpen = defineModel("visible", { default: false }) as Ref<boolean>;
-
-export type DataForTable = { key: string; value: number };
-
-export type PhotoStats = {
-	iso: DataForTable[];
-	focal: DataForTable[];
-	lens: DataForTable[];
-	model: DataForTable[];
-	shutter: DataForTable[];
-	aperture: DataForTable[];
-	year: DataForTable[];
-	month: DataForTable[];
-	day: DataForTable[];
-};
 
 const props = defineProps<{
 	photos: App.Http.Resources.Models.PhotoResource[];
@@ -104,17 +94,11 @@ const props = defineProps<{
 	config: App.Http.Resources.GalleryConfigs.AlbumConfig;
 }>();
 
+const { getStatistics } = useAlbumsStatistics();
+
 const photosData = ref(getStatistics(props.photos));
 const albumSpace = ref(undefined as undefined | App.Http.Resources.Statistics.Album[]);
 const totalAlbumSpace = ref(undefined as undefined | App.Http.Resources.Statistics.Album[]);
-const is_collapsed = ref(false);
-
-const albumData = computed(() => {
-	if (is_collapsed.value === false) {
-		return albumSpace.value?.filter((a) => !a.is_nsfw || are_nsfw_visible.value);
-	}
-	return totalAlbumSpace.value?.filter((a) => !a.is_nsfw || are_nsfw_visible.value);
-});
 
 const total = computed(() => {
 	if (albumSpace.value === undefined) {
@@ -139,74 +123,6 @@ const total = computed(() => {
 
 	return sumData;
 });
-
-function getStatistics(photos: App.Http.Resources.Models.PhotoResource[]): PhotoStats {
-	const stats = {
-		iso: {} as Record<string, number>,
-		focal: {} as Record<string, number>,
-		lens: {} as Record<string, number>,
-		model: {} as Record<string, number>,
-		shutter: {} as Record<string, number>,
-		aperture: {} as Record<string, number>,
-		year: {} as Record<string, number>,
-		month: {} as Record<string, number>,
-		day: {} as Record<string, number>,
-	};
-	for (const photo of photos) {
-		if (photo.precomputed.is_video || photo.precomputed.is_raw) {
-			continue;
-		}
-
-		if (photo.iso) {
-			stats.iso[photo.iso] = stats.iso[photo.iso] ? stats.iso[photo.iso] + 1 : 1;
-		}
-		if (photo.focal) {
-			stats.focal[photo.focal] = stats.focal[photo.focal] ? stats.focal[photo.focal] + 1 : 1;
-		}
-		if (photo.preformatted.aperture) {
-			stats.aperture["ƒ / " + photo.preformatted.aperture] = stats.aperture["ƒ / " + photo.preformatted.aperture]
-				? stats.aperture["ƒ / " + photo.preformatted.aperture] + 1
-				: 1;
-		}
-		if (photo.lens) {
-			stats.lens[photo.lens] = stats.lens[photo.lens] ? stats.lens[photo.lens] + 1 : 1;
-		}
-		if (photo.model) {
-			stats.model[photo.model] = stats.model[photo.model] ? stats.model[photo.model] + 1 : 1;
-		}
-		if (photo.preformatted.shutter) {
-			stats.shutter[photo.preformatted.shutter] = stats.shutter[photo.preformatted.shutter] ? stats.shutter[photo.preformatted.shutter] + 1 : 1;
-		}
-		if (photo.taken_at) {
-			const year = photo.taken_at.slice(0, 4);
-			const month = photo.taken_at.slice(0, 7);
-			const day = photo.taken_at.slice(0, 10);
-			stats.year[year] = stats.year[year] ? stats.year[year] + 1 : 1;
-			stats.month[month] = stats.month[month] ? stats.month[month] + 1 : 1;
-			stats.day[day] = stats.day[day] ? stats.day[day] + 1 : 1;
-		}
-	}
-
-	return {
-		iso: recordToType(stats.iso),
-		focal: recordToType(stats.focal),
-		lens: recordToType(stats.lens),
-		model: recordToType(stats.model),
-		shutter: recordToType(stats.shutter),
-		aperture: recordToType(stats.aperture),
-		year: recordToType(stats.year),
-		month: recordToType(stats.month),
-		day: recordToType(stats.day),
-	};
-}
-
-function recordToType(record: Record<string, number>): DataForTable[] {
-	const data = [] as DataForTable[];
-	Object.entries(record).forEach(([key, value]) => {
-		data.push({ key, value });
-	});
-	return data.sort((a, b) => b.value - a.value);
-}
 
 if (props.config.is_base_album) {
 	StatisticsService.getAlbumSpace(props.album.id).then((response) => {
