+++ resources/js/services/statistics-service.ts
@@ -5,8 +5,8 @@ const StatisticsService = {
 	getUserSpace(): Promise<AxiosResponse<App.Http.Resources.Statistics.UserSpace[]>> {
 		return axios.get(`${Constants.getApiUrl()}Statistics::userSpace`, { data: {} });
 	},
-	getSizeVariantSpace(): Promise<AxiosResponse<App.Http.Resources.Statistics.Sizes[]>> {
-		return axios.get(`${Constants.getApiUrl()}Statistics::sizeVariantSpace`, { data: {} });
+	getSizeVariantSpace(albumId: string | null = null): Promise<AxiosResponse<App.Http.Resources.Statistics.Sizes[]>> {
+		return axios.get(`${Constants.getApiUrl()}Statistics::sizeVariantSpace`, { params: { album_id: albumId }, data: {} });
 	},
 	getAlbumSpace(albumId: string | null = null): Promise<AxiosResponse<App.Http.Resources.Statistics.Album[]>> {
 		return axios.get(`${Constants.getApiUrl()}Statistics::albumSpace`, { params: { album_id: albumId }, data: {} });
